Fixed positioning bugs (jQ 1.12.4 & jQUI 1.12.1)
This might not be the correct way to fix the bugs, but they do work!
Instead of using the JqueryUI Position component only when the user
specifies one, I've changed it to use the Position component as long as
JqueryUI is loaded -- it's likely to be much better at positioning
elements than we are likely to be, given that the Jquery guys probably
know CSS and their module better than we do. :)  In addition, in
situations where JqueryUI is not available, I now check the nodeName
that our div is a child of, because if it's <body> we should call
.offset() and if it's not we should call .position().